minimal_version/requirements.txt
celery==5.3.4
pyyaml==6.0.1
